Refactor chat component structure and styles; consolidate ChatComponent and update imports
