+++ frontend/src/app/app-routing.module.ts
@@ -15,7 +15,7 @@ import {NewWorkoutComponent} from "./features/workouts/components/new-workout/ne
 import {CreateWorkoutComponent} from "./features/workouts/components/create-workout/create-workout.component";
 import {WorkoutsHistoryComponent} from "./features/workouts/components/workouts-history/workouts-history.component";
 import {WorkoutHistoryComponent} from "./features/workouts/components/workout-history/workout-history.component";
-import {ChatComponent} from "./features/chats/components/chat/chat.component";
+import {ChatComponent} from "./features/chats/components/chat.component";
 
 const routes: Routes = [
   {path: '', component: PostListComponent, canActivate: [AuthGuard] },
