+++ frontend/src/app/app.module.ts
@@ -15,7 +15,7 @@ import {AuthInterceptor} from "./core/services/auth.interceptor";
 import {WorkoutsModule} from "./features/workouts/workouts.module";
 import { ChatSidebarComponent } from './features/chats/components/chat-sidebar/chat-sidebar.component';
 import { ChatAreaComponent } from './features/chats/components/chat-area/chat-area.component';
-import { ChatComponent } from './features/chats/components/chat/chat.component';
+import { ChatComponent } from './features/chats/components/chat.component';
 
 @NgModule({
   declarations: [
