+++ frontend/src/app/features/chats/components/chat-sidebar/chat-sidebar.component.scss
@@ -6,4 +6,11 @@
 .chat-sidebar {
   width: 30%;
   background-color: red;
+
+  .search {
+    padding: 10px;
+    margin: 10px;
+    background-color: #C5C6C7;
+    border-radius: 10px;
+  }
 }
