--- frontend/src/app/features/chats/components/chat/chat.component.html
@@ -1,2 +1,0 @@
-<app-chat-sidebar></app-chat-sidebar>
-<app-chat-area></app-chat-area>
