--- frontend/src/app/features/chats/components/chat/chat.component.scss
@@ -1,4 +1,0 @@
-:host {
-  width: 100%;
-  height: 100%;
-}
