--- frontend/src/app/features/chats/components/chat/chat.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-chat',
-  templateUrl: './chat.component.html',
-  styleUrl: './chat.component.scss'
-})
-export class ChatComponent {
-
-}
